+++ modules/flakes.nix
@@ -2,9 +2,7 @@ let
   shared = { config, pkgs, lib, ... }: {
     nix.package = lib.mkDefault pkgs.nix;
     # a < b | a == b
-    nix.settings.experimental-features =
-      assert builtins.compareVersions config.nix.package.version "2.18.1" < 1;
-      "nix-command flakes repl-flake";
+    nix.settings.experimental-features = "nix-command flakes";
     nix.settings.warn-dirty = false;
   };
 in {