+++ overlays/vscode-extensions.nix
@@ -55,18 +55,5 @@ in {
             --replace "wget --no-proxy" "wget --no-proxy --no-continue"
         '';
       });
-  }) (fromMarketplaceRefs [
-    {
-      name = "comment-tagged-templates";
-      publisher = "bierner";
-      version = "0.3.1";
-      sha256 = "sha256-dJyc7txc3fSlNWNGx2G8yF0hObYaiE2c44vzMrvzdkE=";
-    }
-    {
-      name = "markdown-preview-github-styles";
-      publisher = "bierner";
-      version = "1.0.1";
-      sha256 = "sha256-UhWbygrGh0whVxfGcEa+hunrTG/gfHpXYii0E7YhXa4=";
-    }
-  ]);
+  }) (fromMarketplaceRefs [ ]);
 }